// src/models/user/UserInput.ts
type UserInput = {
    email: string;
    password: string;
    confirmPassword: string;
    name: string;
    lastName: string;
    birthDate: Date;
    street: string;
    district: string;
    city: string;
    county: string;
    zipCode: string;
    latitude: string;
    longitude: string;
};

export default UserInput;